pyproject.toml
[project]
name = "drf-api-template"
version = "0.0.1"
description = "Clean template for Django REST Framework applications"
requires-python = ">=3.12"
dependencies = [
    "Django",
    "djangorestframework",
    "drf-spectacular",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "black",
    "mypy",
    "pytest",
    "pytest-django",
    "schemathesis",
    "pre-commit",
    "commitizen",
    "factory-boy",
]

[tool.ruff]
line-length = 88
src = ["src"]
exclude = ["__init__.py",
              ".venv",
              "git",
              "__pycache__",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "N", "C4", "B"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
ignore_missing_imports = true
allow_redefinition = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --import-mode=importlib --capture=sys"
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["src"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
